mtls-proxy/requirements.txt
fastapi==0.103.1
python-dotenv==1.0.0
black==23.9.1
isort==5.12.0
mypy==1.5.1
mypy-extensions==1.0.0
pylint==2.17.6
httpx==0.26.0
uvicorn==0.23.2